split help, command, option
